version 0.5.0: correct working with WP 2.8; added wp-slideshow.js;

+++ trunk/timthumb.php
@@ -15,11 +15,30 @@
 	HTML example: <img src="/scripts/timthumb.php?src=/images/whatever.jpg&w=150&h=200&zc=1" alt="" />
 */
 
-error_reporting(0);
+/*
+$sizeLimits = array(
+	"100x100",
+	"150x150",
+);
+*/
 
-define("CACHE_SIZE", 30);		// number of files to store before clearing cache
-define("CACHE_CLEAR", 5);		// maximum number of files to delete on each cache clear
-define("VERSION", "1.07");		// version number (to force a cache refresh
+define ('CACHE_SIZE', 250);		// number of files to store before clearing cache
+define ('CACHE_CLEAR', 5);		// maximum number of files to delete on each cache clear
+define ('VERSION', '1.09');		// version number (to force a cache refresh
+
+$imageFilters = array(
+	"1" => array(IMG_FILTER_NEGATE, 0),
+	"2" => array(IMG_FILTER_GRAYSCALE, 0),
+	"3" => array(IMG_FILTER_BRIGHTNESS, 1),
+	"4" => array(IMG_FILTER_CONTRAST, 1),
+	"5" => array(IMG_FILTER_COLORIZE, 4),
+	"6" => array(IMG_FILTER_EDGEDETECT, 0),
+	"7" => array(IMG_FILTER_EMBOSS, 0),
+	"8" => array(IMG_FILTER_GAUSSIAN_BLUR, 0),
+	"9" => array(IMG_FILTER_SELECTIVE_BLUR, 0),
+	"10" => array(IMG_FILTER_MEAN_REMOVAL, 0),
+	"11" => array(IMG_FILTER_SMOOTH, 0),
+);
 
 // sort out image source
 $src = get_request("src", "");
@@ -33,10 +52,11 @@ $src = cleanSource($src);
 $lastModified = filemtime($src);
 
 // get properties
-$new_width = preg_replace("/[^0-9]+/", "", get_request("w", 0));
-$new_height = preg_replace("/[^0-9]+/", "", get_request("h", 0));
-$zoom_crop = preg_replace("/[^0-9]+/", "", get_request("zc", 1));
-$quality = preg_replace("/[^0-9]+/", "", get_request("q", 80));
+$new_width 		= preg_replace("/[^0-9]+/", "", get_request("w", 0));
+$new_height 	= preg_replace("/[^0-9]+/", "", get_request("h", 0));
+$zoom_crop 		= preg_replace("/[^0-9]+/", "", get_request("zc", 1));
+$quality 		= preg_replace("/[^0-9]+/", "", get_request("q", 80));
+$filters		= get_request("f", "");
 
 if ($new_width == 0 && $new_height == 0) {
 	$new_width = 100;
@@ -52,6 +72,7 @@ $mime_type = mime_type($src);
 
 // check to see if this image is in the cache already
 check_cache( $cache_dir, $mime_type );
+
 // if not in cache then clear some space and generate a new file
 cleanCache();
 
@@ -78,7 +99,7 @@ if(strlen($src) && file_exists($src)) {
 	// Get original width and height
 	$width = imagesx($image);
 	$height = imagesy($image);
-
+	
 	// don't allow new width or height to be greater than the original
 	if( $new_width > $width ) {
 		$new_width = $width;
@@ -89,15 +110,15 @@ if(strlen($src) && file_exists($src)) {
 
 	// generate new w/h if not provided
 	if( $new_width && !$new_height ) {
-	
+		
 		$new_height = $height * ( $new_width / $width );
 		
 	} elseif($new_height && !$new_width) {
-	
+		
 		$new_width = $width * ( $new_height / $height );
 		
 	} elseif(!$new_width && !$new_height) {
-	
+		
 		$new_width = $width;
 		$new_height = $height;
 		
@@ -114,11 +135,11 @@ if(strlen($src) && file_exists($src)) {
 	imagesavealpha($canvas, true);
 
 	if( $zoom_crop ) {
-	
+
 		$src_x = $src_y = 0;
 		$src_w = $width;
 		$src_h = $height;
-		
+
 		$cmp_x = $width  / $new_width;
 		$cmp_y = $height / $new_height;
 
@@ -145,11 +166,51 @@ if(strlen($src) && file_exists($src)) {
 
 	}
 	
+	if ($filters != "") {
+		// apply filters to image
+		$filterList = explode("|", $filters);
+		foreach($filterList as $fl) {
+			$filterSettings = explode(",", $fl);
+			if(isset($imageFilters[$filterSettings[0]])) {
+			
+				for($i = 0; $i < 4; $i ++) {
+					if(!isset($filterSettings[$i])) {
+						$filterSettings[$i] = null;
+					}
+				}
+				
+				switch($imageFilters[$filterSettings[0]][1]) {
+				
+					case 1:
+					
+						imagefilter($canvas, $imageFilters[$filterSettings[0]][0], $filterSettings[1]);
+						break;
+					
+					case 2:
+					
+						imagefilter($canvas, $imageFilters[$filterSettings[0]][0], $filterSettings[1], $filterSettings[2]);
+						break;
+					
+					case 3:
+					
+						imagefilter($canvas, $imageFilters[$filterSettings[0]][0], $filterSettings[1], $filterSettings[2], $filterSettings[3]);
+						break;
+					
+					default:
+					
+						imagefilter($canvas, $imageFilters[$filterSettings[0]][0]);
+						break;
+						
+				}
+			}
+		}
+	}
+	
 	// output image to browser based on mime type
-	show_image( $mime_type, $canvas, $cache_dir );
+	show_image($mime_type, $canvas, $cache_dir);
 	
 	// remove image from memory
-	imagedestroy( $canvas );
+	imagedestroy($canvas);
 	
 } else {
 
@@ -249,21 +310,35 @@ function cleanCache() {
 
 	$files = glob("cache/*", GLOB_BRACE);
 	
-	usort($files, "filemtime_compare");
+	$yesterday = time() - (24 * 60 * 60);
 	
-	$i = 0;
-	
-	if (count($files) > CACHE_SIZE) {
-		foreach ($files as $file) {
+	if (count($files) > 0) {
 		
-			$i ++;
-			if ($i >= CACHE_CLEAR) {
-				return;
+		usort($files, "filemtime_compare");
+		$i = 0;
+		
+		if (count($files) > CACHE_SIZE) {
+			
+			foreach ($files as $file) {
+				
+				$i ++;
+				
+				if ($i >= CACHE_CLEAR) {
+					return;
+				}
+				
+				if (filemtime($file) > $yesterday) {
+					return;
+				}
+				
+				unlink($file);
+				
 			}
 			
-			unlink($file);
 		}
+		
 	}
+
 }
 
 /**
@@ -276,24 +351,35 @@ function filemtime_compare($a, $b) {
 }
 
 /**
- * 
+ * determine the file mime type
  */
 function mime_type($file) {
 
-    $os = strtolower(php_uname());
+	if (stristr(PHP_OS, 'WIN')) { 
+		$os = 'WIN';
+	} else { 
+		$os = PHP_OS;
+	}
+
 	$mime_type = '';
 
+	if (function_exists('mime_content_type')) {
+		$mime_type = mime_content_type($file);
+	}
+	
 	// use PECL fileinfo to determine mime type
-	if (function_exists('finfo_open')) {
-		$finfo = finfo_open(FILEINFO_MIME);
-		$mime_type = finfo_file($finfo, $file);
-		finfo_close($finfo);
+	if (!valid_src_mime_type($mime_type)) {
+		if (function_exists('finfo_open')) {
+			$finfo = finfo_open(FILEINFO_MIME);
+			$mime_type = finfo_file($finfo, $file);
+			finfo_close($finfo);
+		}
 	}
 
 	// try to determine mime type by using unix file command
 	// this should not be executed on windows
-    if (!valid_src_mime_type($mime_type) && !(eregi('windows', $os))) {
-		if (preg_match("/freebsd|linux/", $os)) {
+    if (!valid_src_mime_type($mime_type) && $os != "WIN") {
+		if (preg_match("/FREEBSD|LINUX/", $os)) {
 			$mime_type = trim(@shell_exec('file -bi "' . $file . '"'));
 		}
 	}
@@ -405,14 +491,11 @@ function show_cache_file($cache_dir) {
 function get_cache_file() {
 
 	global $lastModified;
-
 	static $cache_file;
 	
 	if(!$cache_file) {
-		
-		$cachename = $_SERVER["QUERY_STRING"] . VERSION . $lastModified;
-		$cache_file = md5($cachename) . ".png";
-		
+		$cachename = $_SERVER['QUERY_STRING'] . VERSION . $lastModified;
+		$cache_file = md5($cachename) . '.png';
 	}
 	
 	return $cache_file;
@@ -420,20 +503,20 @@ function get_cache_file() {
 }
 
 /**
- * 
+ * check to if the url is valid or not
  */
 function valid_extension ($ext) {
 
-	if( preg_match( "/jpg|jpeg|png|gif/i", $ext ) ) {
-		return 1;
+	if (preg_match("/jpg|jpeg|png|gif/i", $ext)) {
+		return TRUE;
 	} else {
-		return 0;
+		return FALSE;
 	}
 	
 }
 
 /**
- * 
+ * tidy up the image source url
  */
 function cleanSource($src) {
 
@@ -449,7 +532,7 @@ function cleanSource($src) {
 	$src = str_replace($host, "", $src);
 	$host = str_replace("www.", "", $host);
 	$src = str_replace($host, "", $src);
-	
+
 	// don't allow users the ability to use '../' 
 	// in order to gain access to files below document root
 
@@ -457,9 +540,7 @@ function cleanSource($src) {
 	// src=images/img.jpg or src=/images/img.jpg
 	// not like:
 	// src=../images/img.jpg
-	$src = preg_replace( "/\.\.+\//", "", $src );
-
-	//print_r($_SERVER);
+	$src = preg_replace("/\.\.+\//", "", $src);
 	
 	// get path to image on file system
 	$src = get_document_root($src) . '/' . $src;	
@@ -474,9 +555,19 @@ function cleanSource($src) {
 function get_document_root ($src) {
 
 	// check for unix servers
-	if( @file_exists( $_SERVER['DOCUMENT_ROOT'] . '/' . $src ) ) {
+	if(@file_exists($_SERVER['DOCUMENT_ROOT'] . '/' . $src)) {
 		return $_SERVER['DOCUMENT_ROOT'];
 	}
+	
+	// check from script filename (to get all directories to timthumb location)
+	$parts = array_diff(explode('/', $_SERVER['SCRIPT_FILENAME']), explode('/', $_SERVER['DOCUMENT_ROOT']));
+	$path = $_SERVER['DOCUMENT_ROOT'] . '/';
+	foreach ($parts as $part) {
+		$path .= $part . '/';
+		if (file_exists($path . $src)) {
+			return $path;
+		}
+	}	
 	
 	// the relative paths below are useful if timthumb is moved outside of document root
 	// specifically if installed in wordpress themes like mimbo pro:
@@ -490,34 +581,33 @@ function get_document_root ($src) {
 		"../../../../.."
 	);
 	
-	foreach( $paths as $path ) {
-		if( @file_exists( $path . '/' . $src ) ) {
+	foreach($paths as $path) {
+		if(@file_exists($path . '/' . $src)) {
 			return $path;
 		}
 	}
 	
 	// special check for microsoft servers
 	if(!isset($_SERVER['DOCUMENT_ROOT'])) {
-    	$path = str_replace("/", "\\", $_SERVER["ORIG_PATH_INFO"]);
-    	$path = str_replace($path, "", $_SERVER["SCRIPT_FILENAME"]);
+    	$path = str_replace("/", "\\", $_SERVER['ORIG_PATH_INFO']);
+    	$path = str_replace($path, "", $_SERVER['SCRIPT_FILENAME']);
     	
     	if( @file_exists( $path . '/' . $src ) ) {
     		return $path;
     	}
-	}
+	}	
 	
-	displayError("file not found " . $src);
+	displayError('file not found ' . $src);
 
 }
 
 /**
- *
+ * generic error message
  */
-function displayError($errorString = "") {
+function displayError($errorString = '') {
 
 	header('HTTP/1.1 400 Bad Request');
 	die($errorString);
 	
 }
-
 ?>
